[Topology] Remove TriangleToSplit, store all info directly inside the TriangleSubdivider
